split satellites tracked into separate file
